fix: update cart removed
