+++ components/Cart/CartListing.tsx
@@ -1,9 +1,11 @@
+import { useEffect, useState } from "react";
 import { useSelector } from "react-redux";
 import useAddToCartHook from "../../hooks/CartPageHook/useAddToCart";
 import useFetchCartItems from "../../hooks/CartPageHook/useFetchCartItems";
 import getQuotationPostApi from "../../services/api/cart-apis/get-quotation-api";
 import { CONSTANTS } from "../../services/config/app-config";
 import { get_access_token } from "../../store/slices/auth/token-login-slice";
+import { SelectedFilterLangDataFromStore } from "../../store/slices/general_slices/selected-multilanguage-slice";
 import NoDataFound from "../NoRecordFound";
 import ListViewCard from "./ListViewCard";
 
@@ -16,6 +18,17 @@ function CartListing() {
   const { addToCartItem, cLearCartAPIFunc, RemoveItemCartAPIFunc }: any = useAddToCartHook()
   const TokenFromStore: any = useSelector(get_access_token);
   const { SUMMIT_APP_CONFIG } = CONSTANTS
+  const [selectedMultiLangData, setSelectedMultiLangData] = useState<any>();
+  const SelectedLangDataFromStore: any = useSelector(
+    SelectedFilterLangDataFromStore
+  );
+  useEffect(() => {
+    if (
+      Object.keys(SelectedLangDataFromStore?.selectedLanguageData)?.length > 0
+    ) {
+      setSelectedMultiLangData(SelectedLangDataFromStore?.selectedLanguageData);
+    }
+  }, [SelectedLangDataFromStore]);
   const handleQuotation = async (e: any, quot_id: any) => {
     e.preventDefault();
     const getQuotationInCart = await getQuotationPostApi(
@@ -47,14 +60,13 @@ function CartListing() {
                 }
               >
                 {
-                  // selectedMultiLangData?.request_for_quotation
+                  selectedMultiLangData?.request_for_quotation
                 }
-                Request for quotation
               </button>
-              <button className="btn btn-outline-danger text-decoration-none  px-1" onClick={() => cLearCartAPIFunc(cartListingItems?.name)}>Clear cart</button>
+              <button className="btn btn-outline-danger text-decoration-none  px-1" onClick={() => cLearCartAPIFunc(cartListingItems?.name, setCartListingItems)}>Clear cart</button>
             </div>
           </div>
-          <ListViewCard cartListingItems={cartListingItems} setCartListingItems={setCartListingItems} addToCartItem={addToCartItem} RemoveItemCartAPIFunc={RemoveItemCartAPIFunc} />
+          <ListViewCard cartListingItems={cartListingItems} setCartListingItems={setCartListingItems} addToCartItem={addToCartItem} RemoveItemCartAPIFunc={RemoveItemCartAPIFunc} selectedMultiLangData={selectedMultiLangData} />
         </div>
       ) : (<NoDataFound title={'Your cart is empty !!'} message={'Items added to your cart will show up here'} />)
       }
