+++ components/Navbar/Navbar.tsx
@@ -2,11 +2,12 @@ import useNavbar from '../../hooks/GeneralHooks/useNavbar';
 import WebNavBar from './WebNavBar';
 import { SelectedFilterLangDataFromStore } from '../../store/slices/general_slices/selected-multilanguage-slice';
 import { useSelector } from 'react-redux';
+import { selectCart } from '../../store/slices/cart-slices/cart-local-slice';
 
 function Navbar() {
   const { navbarData, isLoading, errorMessage, selectedCurrencyValue, handleLogoutUser } = useNavbar();
   const { selectedLanguageData }: any = useSelector(SelectedFilterLangDataFromStore);
-  console.log(selectedLanguageData, 'navbar')
+  const cartCount = useSelector(selectCart)?.cartCount
   return (
     <WebNavBar
       navbarData={navbarData}
@@ -15,6 +16,7 @@ function Navbar() {
       selectedCurrencyValue={selectedCurrencyValue}
       handleLogoutUser={handleLogoutUser}
       selectedLanguageData={selectedLanguageData}
+      cartCount={cartCount}
     />
   );
 }
